+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+
+- `InputChips` now immediately converts pasted content to chips instead of waiting for a comma as it does when the user is typing.
+
+### Fixed
+
+- `InputChips` now trims whitespace from each value before calling `validate`.
+
 ## [0.7.16] - 2020-01-30
 
 ### Fixed